+++ src/main/java/de/dafuqs/spectrum/predicate/world/CommandPredicate.java
@@ -1,7 +1,6 @@
-package de.dafuqs.spectrum.recipe.fusion_shrine.world_conditions;
+package de.dafuqs.spectrum.predicate.world;
 
 import com.google.gson.*;
-import de.dafuqs.spectrum.recipe.fusion_shrine.*;
 import net.minecraft.server.*;
 import net.minecraft.server.command.*;
 import net.minecraft.server.world.*;
@@ -9,7 +8,6 @@ import net.minecraft.text.*;
 import net.minecraft.util.math.*;
 
 public class CommandPredicate implements WorldConditionPredicate, CommandOutput {
-	
 	public final String command;
 	
 	public CommandPredicate(String command) {
@@ -17,11 +15,13 @@ public class CommandPredicate implements WorldConditionPredicate, CommandOutput 
 	}
 	
 	public static CommandPredicate fromJson(JsonObject json) {
+        if (json == null || json.isJsonNull()) return (CommandPredicate)ANY;
 		return new CommandPredicate(json.get("command").getAsString());
 	}
 	
 	@Override
 	public boolean test(ServerWorld world, BlockPos pos) {
+		if (this == ANY) return true;
 		MinecraftServer minecraftServer = world.getServer();
 		ServerCommandSource serverCommandSource = new ServerCommandSource(this, Vec3d.ofCenter(pos), Vec2f.ZERO, world, 2, "FusionShrine", world.getBlockState(pos).getBlock().getName(), minecraftServer, null);
 		return minecraftServer.getCommandManager().executeWithPrefix(serverCommandSource, command) > 0;
