+++ src/main/java/de/dafuqs/spectrum/predicate/world/DimensionPredicate.java
@@ -1,7 +1,6 @@
-package de.dafuqs.spectrum.recipe.fusion_shrine.world_conditions;
+package de.dafuqs.spectrum.predicate.world;
 
 import com.google.gson.*;
-import de.dafuqs.spectrum.recipe.fusion_shrine.*;
 import net.minecraft.server.world.*;
 import net.minecraft.util.*;
 import net.minecraft.util.math.*;
@@ -19,6 +18,7 @@ public class DimensionPredicate implements WorldConditionPredicate {
 	}
 	
 	public static DimensionPredicate fromJson(JsonObject json) {
+        if (json == null || json.isJsonNull()) return (DimensionPredicate)ANY;
 		List<RegistryKey<World>> dimensionKeys = new ArrayList<>();
 		for (JsonElement element : json.get("worlds").getAsJsonArray()) {
 			dimensionKeys.add(RegistryKey.of(Registry.WORLD_KEY, Identifier.tryParse(element.getAsString())));
@@ -28,6 +28,7 @@ public class DimensionPredicate implements WorldConditionPredicate {
 	
 	@Override
 	public boolean test(ServerWorld world, BlockPos pos) {
+		if (this == ANY) return true;
 		return this.dimensionKeys.contains(world.getRegistryKey());
 	}
