+++ src/main/java/de/dafuqs/spectrum/predicate/world/WeatherPredicate.java
@@ -1,7 +1,6 @@
-package de.dafuqs.spectrum.recipe.fusion_shrine.world_conditions;
+package de.dafuqs.spectrum.predicate.world;
 
 import com.google.gson.*;
-import de.dafuqs.spectrum.recipe.fusion_shrine.*;
 import net.minecraft.server.world.*;
 import net.minecraft.util.math.*;
 
@@ -24,11 +23,13 @@ public class WeatherPredicate implements WorldConditionPredicate {
 	}
 	
 	public static WeatherPredicate fromJson(JsonObject json) {
+        if (json == null || json.isJsonNull()) return (WeatherPredicate)ANY;
 		return new WeatherPredicate(WeatherCondition.valueOf(json.get("weather_condition").getAsString().toUpperCase(Locale.ROOT)));
 	}
 	
 	@Override
 	public boolean test(ServerWorld world, BlockPos pos) {
+		if (this == ANY) return true;
 		switch (this.weatherCondition) {
 			case CLEAR_SKY -> {
 				return !world.isRaining();
