+++ src/main/java/de/dafuqs/spectrum/recipe/fusion_shrine/FusionShrineRecipe.java
@@ -5,6 +5,7 @@ import de.dafuqs.spectrum.blocks.*;
 import de.dafuqs.spectrum.blocks.fusion_shrine.*;
 import de.dafuqs.spectrum.blocks.upgrade.*;
 import de.dafuqs.spectrum.helpers.*;
+import de.dafuqs.spectrum.predicate.world.WorldConditionPredicate;
 import de.dafuqs.spectrum.recipe.*;
 import de.dafuqs.spectrum.registries.*;
 import net.id.incubus_core.recipe.*;
