+++ src/main/java/de/dafuqs/spectrum/recipe/fusion_shrine/FusionShrineRecipeSerializer.java
@@ -1,6 +1,8 @@
 package de.dafuqs.spectrum.recipe.fusion_shrine;
 
 import com.google.gson.*;
+
+import de.dafuqs.spectrum.predicate.world.WorldConditionPredicate;
 import de.dafuqs.spectrum.recipe.*;
 import net.id.incubus_core.recipe.*;
 import net.minecraft.fluid.*;
